requirements.txt
python-cowsay>=1.1.1
